close input stream when all infile data is written
there is some problems that contains a several test cases, and need to
read until we come across EOF. but rust's try_wait() doesn't close input
stream (while wait() does but for TLE we can't use this) .

the solution is dropping ChildStdin from ChildProcess after all of input
data is written. we cannot move out of child because child will be a
partially moved data so we can't use it after that. instead, it can be
done with Option's take() method.