fix minify bug
before this, because of poor treatment of newline, the minify process
causes compile error even if it compiles and runs well in run command.
for example:

```
if (true)
    process();
else
    other_process();
```

this was minified to:

```
if(true)process();elseother_process();
               // ^^^^^^^^^ two words are joined unexpectedly
```

to avoid this, now it insert space (' ') between lines. in other words,
above example was first converted to the next ...

```
if (true) process; else other_process();
```

... and then the normal minify procedure starts. at last, we got below:

```
if(true)process;else other_process();
```