init: change to create new directory
previously init command initializes project in a current directory, but
most of the use case is to create new empty directory for problem and
`cd` to the directory, and then run `procon-assistant init` in the
destination directory. this is very complex style.

so now init command takes one argument --- the directory name. now init
command creates new directory named specified, and then initialize new
project in it.