init: use "." as a project name by default
the enhancement in init command specifying project name and create
directory for the project, is very useful for a single task, but for the
already initialized directory, such as generated by initdirs command,
needs extra argument "." to init the project. this is not useful, so, if
the project name was not specified, i.e. called with `pa i` or else,
then use "." as a default project name.