add `-fno-delayed-template-expansion` in Windows
In LLVM (clang) for Windows, some compiler options are defaulted to
stuffs keeping compatibility with Visual Studio. Option
`-fdelayed-template-expansion` is one of them, for already written and
working well with Visual Studio's way of non two-phase argument lookup
aware template processing. If this option is enabled many errors cannot
be found in local machine, especially forgetting adding `std::` is, if
`using namespace std;` is present *AFTER* that template function, not an
error. In other words, below example is successfully compiled:

    #include <iostream>
    template <typename T> T cget() {
        T x; cin >> x; return x;
        //   ^^^ MUST be std::cin based on C++ Standard.
    }

    // this using declaration must not affect lines before this
    // declaration; yes, cget() is out of scope.
    using namespace std;
    int main() {}

This however compilation error in gcc/clang in Linux (again, this is
correct behavior). So if this code is submitted, it causes CE
(Compilation Error) in judge. To disable this troublesome behavior, I
add `-fno-delayed-template-expansion` for Windows.