do not create root directory for local contest
Consider when the `problems.txt` is in current directory and run
`procon-assistant download local:problems.txt`. With the previous
implementation, the directory named `Local` is created in the current
directory, and the each problem directories are created under the
`Local` directory, as follows:

```
(current dir)
|- problems.txt
\- Local
   |- a
   |- b
      :
```

However the user likely expects problem directories created in the
current directory, as follows:

```
(current dir)
|- problems.txt
|- a
|- b
   :
```

This commit adds `unique_contest_id` flag and the contest directory will
not created when it's true.

Fix #15