support large output
Previously output is read after waiting the process to finish. However,
 this causes a critical issue: if the stdout / stderr's buffer becomes
full, then the program's execution is blocked and we have deadlock.
To prevent this, read the child output from another thread during
execution. Normally `read_to_string()` before `wait()` is enough, but
this time another thread is required because we must have timeout
to the execution.